others code

+++ src/component/Admin/Delete/Delete.js
@@ -7,13 +7,17 @@ const Delete = ({dele}) => {
     const [deleteCourse,setDeleteCourse]=useState([])
                 
     const deleteProduct = (id) => {
-        console.log(id);
+        // console.log(id);
+
+        const proceed = window.confirm("Are you sure, you want to delete?");
+  if (proceed) {
         fetch(`http://localhost:5000/delete/${id}`, {
           method: 'DELETE'
         })
         .then(res => res.json())
         .then(result => setDeleteCourse(result))
         Swal.fire("Good job!", "Data Deleted Successfully!");
+      }
         }
 
 
@@ -21,26 +25,27 @@ const Delete = ({dele}) => {
     
 
     return (
-        <div>
-           
-           
-           <table style={{width:"200px"}} class="table">
-  <thead class="thead-dark">
-    <tr style={{width:"200px"}}>
-      <th scope="col">Name</th>
-      <th scope="col">Button</th>
-    </tr>
-  </thead>
   
-  <tbody>
-    <tr>
-      <th scope="row">{dele.name}</th>
-      <td><button onClick={() => deleteProduct(dele._id)} className="btn btn-primary" variant="primary">Delete</button></td>
-    </tr>
-    
-  </tbody>
-</table>
-        </div>
+
+        <div className="col-md-4 text-center pt-5 ">
+      
+      <img
+        className="img"
+        style={{ height: "100px" }}
+        src={dele?.imageURL}
+        alt=""
+      />
+      <h4 className="pt-3">{dele?.name}</h4>
+     
+    <div className="d-flex p-2">
+    <div className="mx-1">
+    <button onClick={() => deleteProduct(dele?._id)} className="btn btn-danger" variant="primary">Delete </button>
+     
+    </div>
+    <div className="mx-1"></div>
+     
+      </div>
+    </div>
     );
 };
